backend/requirements.txt
Django>=5.0
djangorestframework
psycopg2-binary
gunicorn
django-environ
djangorestframework-simplejwt
django-redis
django-cors-headers